+++ provisioners/aws/rhel7/post_puppet.sh
@@ -1,9 +1,0 @@
-#!/usr/bin/env bash
-set -o nounset
-set -o errexit
-
-# Clean up temporary files
-rm -rf /tmp/shinesolutions/packer-aem
-
-# Clear history
-history -c