+++ provisioners/bash/noop.sh
@@ -1,0 +1,5 @@
+#!/usr/bin/env bash
+set -o nounset
+set -o errexit
+
+exit 0