+++ provisioners/puppet/modules/config/spec/classes/aem_cleanup_spec.rb
@@ -1,0 +1,3 @@
+require 'spec_helper'
+# This class is only used from inside the ::config::aem class and relies on the
+# Exec defaults set there. Because of that, it is not tested here.