+++ test/serverspec/spec/aem_base_spec.rb
@@ -5,5 +5,5 @@ describe package('nokogiri') do
 end
 
 describe package('ruby_aem') do
-  it { should be_installed.by('gem').with_version('1.3.0') }
+  it { should be_installed.by('gem').with_version('1.3.1') }
 end