+++ test/serverspec/spec/soe_spec.rb
@@ -1,0 +1,3 @@
+require 'spec_helper'
+
+# TODO Implement some tests.