+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+### 2.1.0
+* Service names are now 'aem-author' (AEM Author) and 'aem-publish' (AEM Publish), previously 'aem-aem'
+* Replace default https port for author component from 5433 to 5432
+* Added AuthorPublishDispatcher component
+* Added RHEL 7.x support
+* Replace librarian-puppet with r10k for Puppet dependency management
+
+### 2.0.0
+* (NEED TO COMPLETE)
+
 ### 1.1.0
 * Upgrade AEM Cumulative Fix Pack to SP1-CFP3
 * Update to Java Version 8u131